frontend/requirements.txt
streamlit
requests
google-auth>=2.15.0
cryptography